+++ .github/release-notes.md
@@ -1,2 +1,3 @@
 * Fixed players not being properly sorted after reloading
 * Fixed last player to join sometimes showing older profile picture
+* Added button to scroll to top on history pages